move planning screen into folder where it belongs
